Created library to manage the Arcade Buttons
Each button is assigned a number and description, these variables could help facilitate debugging of the arcade unit as a whole. Each button is also assigned a Digital Pin, as they are buttons each pin will be set to INPUT.

When you construct the ArcadeButton object it will also trigger init() which will set the initial value of the button state to `UnPressed`. It will not matter if the button is actually being depressed when the system comes online because it will automatically correct itself. To manage the button states I have built a miniature State Machine within TransitionButtonState(), the rules for transitions are stored within this function.