Added support for ASCII Key Codes and additional getter methods
I wanted the button to understand what it should send to the keyboard once that portion of the sketch is working, so to support this I have added the ASCIIKeyCode variable in the constructor. Also added more getter methods so that it is now possible to get all of the values from the constructor back.