Added overload for the constructor
Not all of the buttons on an arcade would be sent back to the computer via keyboard key code. For example X-Arcade's Tank Stick has a programming button that is not sent to the PC. In addition I resolved a compilation error as `string`s do not exist, but instead they are `String`s.